Slight re-org & protect account edit when require_mfa is true
