+++ app/controllers/application_controller.rb
@@ -7,21 +7,9 @@ class ApplicationController < ActionController::Base
 
   include Pwpush::FirstRun
   include SetLocale
-
-  before_action :enforce_mfa_if_required
+  include EnforceRequiredMfa
 
   private
-
-  def enforce_mfa_if_required
-    return unless user_signed_in?
-    return unless Settings.require_mfa
-    return if current_user.otp_required_for_login?
-    return if controller_path == "users/two_factor"
-    return if controller_path == "users/sessions"
-
-    redirect_to backup_codes_user_two_factor_path,
-      alert: _("Two-factor authentication is required. Please set it up to continue.")
-  end
 
   # To add extra fields to Devise registration, add the attribute names to `extra_keys`
   # See: https://stackoverflow.com/questions/64057147/attributes-not-saving-with-devise-and-accepts-nested-attributes-for
