+++ app/controllers/users/registrations_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Users::RegistrationsController < Devise::RegistrationsController
+  include EnforceRequiredMfa
+
   invisible_captcha only: :create
 
   layout "application", only: %i[edit update token]
